ussd_gateway/ussd_gateway/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==1.10.4
python-multipart
requests
psycopg2-binary==2.9.9
